+++ Tweaks/UiAdjustment/ImprovedInterruptableCastbars.cs
@@ -118,7 +118,7 @@ public unsafe class ImprovedInterruptableCastbars : UiAdjustments.SubTweak{
         var imageNode = UiHelper.MakeImageNode(nodeId, new UiHelper.PartInfo(0, 0, 36, 36));
         imageNode->NodeFlags = NodeFlags.AnchorLeft | NodeFlags.AnchorTop | NodeFlags.Visible | NodeFlags.Enabled | NodeFlags.EmitsEvents;
         imageNode->WrapMode = 1;
-        imageNode->Flags = (byte) ImageNodeFlags.AutoFit;
+        imageNode->Flags = ImageNodeFlags.AutoFit;
         
         imageNode->LoadIconTexture((uint)icon, 0);
         imageNode->ToggleVisibility(true);
